Find the root folder in a slightly better way.
HaxeComplete used to assume that the root folder of the project was the
folder that the current file was in, which is not generally true. The code
here works for the case where you only have one root folder open in
Sublime - it assume that's the root folder and searches for *.nmml files
in it. If you have more than one, it reverts to the old way of looking
around where your file is, which is probably what you intended if you have
multiple folders. However this could still be improved.