Show documentation in separate view.
This will open up a new view in Sublime Text and, as you type a function
name, it'll take a best guess at what you're trying to type and show you
the documentation for that function.

Some notes about this.

1. There's no way to correctly know what the user has selected in the
autocomplete box, as far as I can tell. That's why we have to take a
guess in HaxeShowDocumentation#run.

2. This automatically assumes that any window in the second row of
Sublime Text is the documentation. This is not necessarily true.

3. The first time it opens the window, if you're using some sort of
vi-emulation like Vintage or Vintageous, it'll snap you out of insert
mode. Since it's only the first time (and then never again) I haven't
invested the time to figure this one out. Plus, it's not a big issue.

4. It would be nice to have an option to toggle this off. Currently, if
you don't like it, that's too bad, because if you close it, it'll just
open up again. (Though you can make it as small as you want.)