Fix bug where searching would trigger documentation view.
This is a slight hack, but I can't find any way to directly ask Sublime
if the user's input is currently in a search bar.

I also changed the function from on_selection_modified_async to
on_modified_async, because this function is only called when the user
types in something. Previously, the function would be called even if the
user moved the selection around with the arrow keys without typing.