Fix bug with documentation.
Unfortunately, calling sel() on active_view() will just return the same
object every time, so the comparison I was doing would always fail.
There's no easy way to compare sel() objects - you'd have to loop
through every selection to ensure it's right, and it gets even more
painful to duplicate - so I decided to check if the first region has
changed to see if we're typing in the find window or not.

If the user is actually using multiple selections, it's fine,
since it's hard to say which of the functions they're selecting should
be shown documentation anyways.